Refactor form fields into separate components
