+++ src/dashboard/form-manager/components/CustomField.tsx
@@ -1,0 +1,18 @@
+const CustomField = ( {
+    label,
+    children,
+    className = '',
+}: {
+    label: string | React.ReactNode;
+    children: React.ReactNode;
+    className?: string;
+} ) => {
+    return (
+        <div className={ `flex flex-col gap-1 ${ className }` }>
+            <div className="uppercase">{ label }</div>
+            { children }
+        </div>
+    );
+};
+
+export default CustomField;
